Changed type of 'offset' parameter to use int instead Removed unused NagerDateHttpClient parameter from Context.java Remove unused parameter 'key' Fixed ignored --version flag Fixed GetLongWeekend ignores --available-bridge-days and --subdivision Updated README.md to reflect Java 21 requirement and removed --key option

+++ nager-date-http-client/src/main/java/com/github/flombois/services/v3/HttpPublicHolidayV3Service.java
@@ -74,7 +74,7 @@ public class HttpPublicHolidayV3Service implements PublicHolidayV3Service {
 
     /** {@inheritDoc} */
     @Override
-    public boolean isTodayAPublicHoliday(CountryCode countryCode, String countyCode, byte offset) throws NagerDateServiceException {
+    public boolean isTodayAPublicHoliday(CountryCode countryCode, String countyCode, int offset) throws NagerDateServiceException {
         try {
             return client.callApi(getIsTodayPublicHolidayUri(countryCode, countyCode, offset), new ResponseHandler<>() {
                 @Override
@@ -101,7 +101,7 @@ public class HttpPublicHolidayV3Service implements PublicHolidayV3Service {
      * @param offset the timezone offset from UTC in hours
      * @return the constructed URI path with query parameters
      */
-    protected String getIsTodayPublicHolidayUri(CountryCode countryCode, String countyCode, byte offset) {
+    protected String getIsTodayPublicHolidayUri(CountryCode countryCode, String countyCode, int offset) {
         return IS_TODAY_PUBLIC_HOLIDAY_ENDPOINT + "/" + countryCode.name() + toQueryParam(countyCode, offset);
     }
